refact(settler): separate settler entity into repository and service
- in order to enhance the organization, I am separating this structure.
Now `Services` classes will provide services' methods to external
components, while Repositories will actually manipulate entities in a
world.
