+++ Assets/Scripts/Game/Unit/Settler/Repository.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a67bec485416e364fb10ff3d5e6af6d5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
