+++ Assets/Scripts/Game/Unit/Settler/Repository/SettlerRepository.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d88e142e6d20cb94a9aa2217ebcac986
+guid: 5aff8adcdc0ef344f8b718f2902fb5ff
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
