+++ Assets/Scripts/Game/Unit/Settler/Services.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e89d446a51b96b49b25b7b6e3801e8e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
