+++ Assets/Scripts/Game/Unit/Settler/Services/SettlerService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bac4c3c41be864148b056a29a76e1d2c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
